Register music assistant services in async setup
